pyproject.toml
[project]
name = "pyinstaller-github-action"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=11.0.0",
    "pyinstaller>=6.11.1",
]
